fix: do not create main index file if one present

+++ src/src/src.ts
@@ -2,21 +2,27 @@ import Generator from 'yeoman-generator';
 import { dependencies } from '../core';
 
 export default class SrcGenerator extends Generator {
+  private mainPath: string = '';
+
   public initializing(): void {
-    dependencies.importFromPkg(this.fs.readJSON('package.json'));
+    const pkg = this.fs.readJSON('package.json');
+    dependencies.importFromPkg(pkg);
+    this.mainPath = this.destinationPath(pkg.main);
   }
 
   public writing(): void {
-    let extension = 'js';
+    if (!this.fs.exists(this.mainPath)) {
+      let extension = 'js';
 
-    if (dependencies.has('typescript', 'devDependencies')) {
-      extension = 'ts';
+      if (dependencies.has('typescript', 'devDependencies')) {
+        extension = 'ts';
+      }
+
+      if (dependencies.has('react')) {
+        extension += 'x';
+      }
+
+      this.fs.write(this.destinationPath(`src/index.${extension}`), '');
     }
-
-    if (dependencies.has('react')) {
-      extension += 'x';
-    }
-
-    this.fs.write(this.destinationPath(`src/index.${extension}`), '');
   }
 }
